+++ cobaya/likelihoods/base_classes/InstallableLikelihood.py
@@ -16,7 +16,7 @@ from cobaya.typing import InfoDict
 from cobaya.log import get_logger
 from cobaya.install import _version_filename
 from cobaya.component import ComponentNotInstalledError
-from cobaya.tools import VersionCheckError
+from cobaya.tools import VersionCheckError, resolve_packages_path
 
 
 class InstallableLikelihood(Likelihood):
@@ -30,9 +30,13 @@ class InstallableLikelihood(Likelihood):
         # Ensure check for install and version errors
         # (e.g. may inherit from a class that inherits from this one, and not have them)
         if self.install_options:
-            if not self.is_installed(path=kwargs["packages_path"]):
+            packages_path = kwargs.get("packages_path") or resolve_packages_path()
+            if not self.is_installed(path=packages_path):
+                name = self.get_qualified_class_name()
+                logger = get_logger(name)
                 raise ComponentNotInstalledError(
-                    "The data for this likelihood has not been correctly installed.")
+                    logger, ("The data for this likelihood has not been correctly "
+                             f"installed. To install it, run `cobaya-install {name}`"))
         super().__init__(*args, **kwargs)
 
     @classmethod
@@ -78,7 +82,7 @@ class InstallableLikelihood(Likelihood):
                 return True
             elif not (os.path.exists(path) and len(os.listdir(path)) > 0):
                 log = get_logger(cls.get_qualified_class_name())
-                log.error ("The given installation path does not exist: '%s'", path)
+                log.error("The given installation path does not exist: '%s'", path)
                 return False
             elif opts.get("github_release"):
                 try: